added all lst and modified part 1 and 2

+++ libft/ft_strlcat.c
@@ -6,17 +6,17 @@
 /*   By: xquah <xquah@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/17 12:12:18 by xquah             #+#    #+#             */
-/*   Updated: 2024/02/23 15:55:12 by xquah            ###   ########.fr       */
+/*   Updated: 2024/03/06 21:49:24 by xquah            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, char *src, unsigned int size)
 {
-	unsigned int	i;
-	unsigned int	j;
-	unsigned int	initial_dest_length;
+	size_t	i;
+	size_t	j;
+	size_t	initial_dest_length;
 
 	initial_dest_length = ft_strlen(dest);
 	if (size <= initial_dest_length)
